finish response fragment

+++ app/src/main/java/com/lnsergioantonio/restapp/domain/model/ResponseEntity.kt
@@ -7,7 +7,7 @@ data class ResponseEntity(
         val url: String,
         val time: String,
         val statusCode: Int,
-        val size: Int,
+        val size: Long,
         val requestBody: String,
         val responseBody: String,
         val isSuccessful: Boolean,
